fuse specifications
