--- specification_templates/fit.yaml
@@ -1,13 +1,0 @@
-data:
-  infection_version: '2020_06_28.02'
-  location_set_version_id: ''
-  output_root: ''
-  location_set_file: '/ihme/covid-19/seir-pipeline-outputs/metadata-inputs/location_metadata_727001.csv'
-parameters:
-  n_draws: 10
-  day_shift: [0, 8]
-  alpha: [0.9, 1.0]
-  sigma: [0.2, 0.3333]
-  gamma1: [0.5, 0.5]
-  gamma2: [0.3333, 1.0]
-  solver_dt: 0.1
