+++ src/covid_model_seiir_pipeline/regression/specification.py
@@ -8,8 +8,28 @@ from covid_model_seiir_pipeline.utilities import Specification, asdict
 class RegressionData:
     """Specifies the inputs and outputs for a regression"""
     covariate_version: str = field(default='best')
-    ode_fit_version: str = field(default='best')
+    infection_version: str = field(default='best')
+    location_set_version_id: int = field(default=0)
+    location_set_file: str = field(default='')
     output_root: str = field(default='')
+
+    def to_dict(self) -> Dict:
+        """Converts to a dict, coercing list-like items to lists."""
+        return asdict(self)
+
+
+@dataclass
+class FitParameters:
+    """Specifies the parameters of the ODE fit."""
+    n_draws: int = field(default=1000)
+
+    day_shift: Tuple[int, int] = field(default=(0, 8))
+
+    alpha: Tuple[float, float] = field(default=(0.9, 1.0))
+    sigma: Tuple[float, float] = field(default=(0.2, 1/3))
+    gamma1: Tuple[float, float] = field(default=(0.5, 0.5))
+    gamma2: Tuple[float, float] = field(default=(1/3, 1.0))
+    solver_dt: float = field(default=0.1)
 
     def to_dict(self) -> Dict:
         """Converts to a dict, coercing list-like items to lists."""
@@ -43,14 +63,17 @@ class RegressionSpecification(Specification):
 
     def __init__(self,
                  data: RegressionData,
+                 parameters: FitParameters,
                  covariates: List[CovariateSpecification]):
         self._data = data
+        self._parameters = parameters
         self._covariates = {c.name: c for c in covariates}
 
     @classmethod
     def parse_spec_dict(cls, regression_spec_dict: Dict) -> Tuple:
         """Constructs a regression specification from a dictionary."""
         data = RegressionData(**regression_spec_dict.get('data', {}))
+        parameters = FitParameters(**regression_spec_dict.get('parameters', {}))
 
         # covariates
         cov_dicts = regression_spec_dict.get('covariates', {})
@@ -66,6 +89,11 @@ class RegressionSpecification(Specification):
     def data(self) -> RegressionData:
         """The data specification for the regression."""
         return self._data
+
+    @property
+    def parameters(self) -> FitParameters:
+        """The parametrization of the regression."""
+        return self._parameters
 
     @property
     def covariates(self) -> Dict[str, CovariateSpecification]:
